+++ src/43rollup.js
@@ -18,14 +18,14 @@ var rollup = function (a,query) {
 		var ss = [];
 		for(var i=0;i<glen;i++) {
 		 	if(a[i] instanceof yy.Column) {
-				a[i].nick = a[i].columnid;
+				a[i].nick = escapeq(a[i].columnid);
 
-		 		query.groupColumns[a[i].columnid] = a[i].nick;
+		 		query.groupColumns[escapeq(a[i].columnid)] = a[i].nick;
 				var aaa = a[i].nick+'\t'
 					+a[i].toJavaScript('p',query.sources[0].alias,query.defcols);
 		 	} else {
-		 		query.groupColumns[a[i].toString()] = a[i].toString();
-				var aaa = a[i].toString()+'\t'
+		 		query.groupColumns[escapeq(a[i].toString())] = escapeq(a[i].toString());
+				var aaa = escapeq(a[i].toString())+'\t'
 					+a[i].toJavaScript('p',query.sources[0].alias,query.defcols);
 			}
 
@@ -93,13 +93,13 @@ function decartes(gv,query) {
 		for(var t=0; t<gv.length; t++) {
 			if(gv[t] instanceof yy.Column) {
 			//	console.log('+++',gv[t].columnid,gv[t]);
-				gv[t].nick = gv[t].columnid;
-			 	query.groupColumns[gv[t].columnid] = gv[t].nick;
+				gv[t].nick = escapeq(gv[t].columnid);
+			 	query.groupColumns[gv[t].nick] = gv[t].nick;
 		 		res = res.map(function(r){return r.concat(gv[t].nick+'\t'+gv[t].toJavaScript('p',query.sources[0].alias,query.defcols))}); 	
 //		 		res = res.map(function(r){return r.concat(gv[t].columnid)}); 	
 			} else if(gv[t] instanceof yy.FuncValue) {
-				query.groupColumns[gv[t].toString()] = gv[t].toString();
-		 		res = res.map(function(r){return r.concat(gv[t].toString()+'\t'+gv[t].toJavaScript('p',query.sources[0].alias,query.defcols))}); 	
+				query.groupColumns[escapeq(gv[t].toString())] = escapeq(gv[t].toString());
+		 		res = res.map(function(r){return r.concat(escapeq(gv[t].toString())+'\t'+gv[t].toJavaScript('p',query.sources[0].alias,query.defcols))}); 	
 		 		// to be defined
 			} else if(gv[t] instanceof yy.GroupExpression) {
 				if(gv[t].type == 'ROLLUP') res = cartes(res,rollup(gv[t].group,query));
@@ -115,8 +115,8 @@ function decartes(gv,query) {
 //				console.log(gv[t].toString());
 //console.log('+++');
 		 		res = res.map(function(r){
- 					query.groupColumns[gv[t].toString()] = gv[t].toString();
-		 			return r.concat(gv[t].toString()
+ 					query.groupColumns[escapeq(gv[t].toString())] = escapeq(gv[t].toString());
+		 			return r.concat(escapeq(gv[t].toString())
 		 				+'\t'
 		 				+gv[t].toJavaScript('p',query.sources[0].alias,query.defcols)) 
 		 		}); 	
@@ -137,17 +137,17 @@ function decartes(gv,query) {
 		return res;
 	} else if(gv instanceof yy.FuncValue) {
 //		console.log(gv);
-		query.groupColumns[gv.toString()] = gv.toString();
+		query.groupColumns[escapeq(gv.toString())] = escapeq(gv.toString());
 		return [gv.toString()+'\t'+gv.toJavaScript('p',query.sources[0].alias,query.defcols)];
 	} else if(gv instanceof yy.Column) {
-			gv.nick = gv.columnid;
-		 	query.groupColumns[gv.columnid] = gv.nick;
+			gv.nick = escapeq(gv.columnid);
+		 	query.groupColumns[gv.nick] = gv.nick;
 			return [gv.nick+'\t'+gv.toJavaScript('p',query.sources[0].alias,query.defcols)]; // Is this ever happened?
 		// } else if(gv instanceof yy.Expression) {
 		// 	return [gv.columnid]; // Is this ever happened?
 	} else {
-		query.groupColumns[gv.toString()] = gv.toString();
-		return [gv.toString()+'\t'+gv.toJavaScript('p',query.sources[0].alias,query.defcols)];
+		query.groupColumns[escapeq(gv.toString())] = escapeq(gv.toString());
+		return [escapeq(gv.toString())+'\t'+gv.toJavaScript('p',query.sources[0].alias,query.defcols)];
 //			throw new Error('Single argument in the group without array');			
 	};