+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 * SET ERRORLOG ON/OFF (trap errors)
 * COALESCE() function skips NaN values as well as NULL
 * SET CACHE ON/OFF - turn on/off SQL statements caching
-* loadFileFrom('#selector') - to read data from <xmp> tag
+* loadFileFrom('#selector') - to read data from HTML <tag> 
 
 ### 0.0.38 "Elbrus" (17.01.2015 - 17.01.2015)